+++ src/UnlimitedGPT/UnlimitedGPT.py
@@ -274,12 +274,71 @@ class ChatGPT:
         self.driver.close()
         self.driver.switch_to.window(original_window)
 
+    def _continue_generating(self, timeout: int) -> Optional[str]:
+        """
+        Continue generating the response.
+
+        Args:
+        ----------
+            timeout (int): Time to wait for the message to continue regenerating before timing out.
+
+        Returns:
+        ----------
+            Optional[str]: The newly generated response.
+        """
+        self.logger.debug('Continuing generating...')
+        # Click "Continue generating" button
+        continue_response_clicked = self.driver.safe_click(
+            CGPTV.continue_regenerating, timeout = 5
+        )
+        if not continue_response_clicked:
+            self.logger.debug('Could not click continue regenerating button')
+            return None
+
+        # Get the response, same way as send_message without the part of sending the message
+        self.logger.debug('Waiting for completion...')
+        try:
+            WebDriverWait(self.driver, timeout).until_not(
+                EC.presence_of_element_located(CGPTV.streaming)
+            )
+        except:
+            self.logger.debug('Could not continue generating')
+            return None
+
+        self.logger.debug('Getting response...')
+        responses = self.driver.find_elements(*CGPTV.big_response)
+        if responses:
+            response = responses[-1]
+            if 'text-red' in response.get_attribute('class'):
+                self.logger.debug('Response is an error')
+                return None
+        response = self.driver.find_elements(*CGPTV.small_response)
+        try:
+            response = response[-1]
+        except IndexError:
+            self.logger.debug('Response not found, resetting conversation...')
+            self.reset_conversation()
+            return None
+        
+        content = markdownify(
+            response.get_attribute('innerHTML'),
+            escape_asterisks=False,
+            escape_underscores=False,
+
+        ).replace(
+            'Copy code`', '`'
+        ).rstrip("\n")
+        
+        self.logger.debug('Continued regenerating the response')
+        return content
+
     def send_message(
         self,
         message: str,
         timeout: int = 240,
         input_mode: Literal['INSTANT', 'SLOW'] = "INSTANT",
         input_delay: float = 0.1,
+        continue_generating: bool = True,
     ) -> ChatGPTResponse:
         """
         Send a message to ChatGPT.
@@ -290,6 +349,7 @@ class ChatGPT:
             timeout (int, optional): Timeout in seconds. Defaults to 240.
             input_mode(list, optional): The input mode. Defaults to 'INSTANT'.
             input_delay(float, optional): The input delay. Defaults to 0.1.
+            continue_generating(bool, optional): Whether to continue generating the response or not. Defaults to True.
 
         Returns:
         ----------
@@ -352,19 +412,33 @@ class ChatGPT:
             'Copy code`', '`'
         ).rstrip("\n")
 
+        if continue_generating:
+            continuation = self._continue_generating(timeout = timeout)
+            if continuation:
+                content = continuation
+
         self.logger.debug(f'Message sent')
 
         return ChatGPTResponse(content, self._conversation_id)
 
-    def regenerate_response(self, message_timeout: int = 240, click_timeout: int = 20) -> ChatGPTResponse:
+    def regenerate_response(
+        self,
+        message_timeout: int = 240,
+        click_timeout: int = 20,
+        continue_generating: bool = True
+    ) -> ChatGPTResponse:
         """
         Regenerate the response.
+
+        Args:
+        ----------
+            message_timeout (int, optional): Time to wait for the message to regenerate before timing out. Defaults to 240.
+            click_timeout (int, optional): Time to wait for the click to succeed before timing out. Defaults to 20.
+            continue_generating(bool, optional): Whether to continue generating the response or not. Defaults to True.
 
         Returns:
         ----------
             response (ChatGPTResponse): The newly regenerated response data.
-            message_timeout (int, optional): Time to wait for the message to regenerate before timing out. Defaults to 240.
-            click_timeout (int, optional): Time to wait for the click to succeed before timing out. Defaults to 20.
 
         Raises:
         ----------
@@ -412,6 +486,11 @@ class ChatGPT:
         ).replace(
             'Copy code`', '`'
         ).rstrip("\n")
+
+        if continue_generating:
+            continuation = self._continue_generating(timeout = message_timeout)
+            if continuation:
+                content = continuation
         
         self.logger.debug('Regenerated response')
         return ChatGPTResponse(content, self._conversation_id)